MPI Monte Carlo pi approximation calculator
MPI or Message Passing Interface is one of the most-common HPC (High
Performance Computing, i.e. Supercomputer) APIs for C/C++ and Fortran.
It handles the parallelization of otherwise serial programs and the
asyncronous calculation across all computation nodes (i.e. CPU cores/HTT
threads of all MPI-enabled hosts in the environment). This is a short
MPI-based program to calculate an approximation of pi using the Monte
Carlo algorithm, which gives a rough approximation of abt 93% accurate
pi.